backend/requirements.txt
asgiref==3.7.2
Django==4.2.3
django-cors-headers==4.3.0
django-enumfields==2.1.1
djangorestframework==3.14.0
dnspython==2.5.0
Pillow==10.0.0
pymongo==4.6.1
python-dotenv==1.0.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
